requirements.txt
asgiref==3.6.0
Django==4.2
django-appconf==1.0.5
django-compressor==4.3.1
django-environ==0.10.0
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.4
pytz==2023.3
djangorestframework==3.14.0

